rename: ThemeButton 기준 파일명 index로 변경
