+++ src/components/common/ThemeButton/index.tsx
@@ -1,7 +1,7 @@
 import SunIcon from '../icons/SunIcon';
 import { useDarkMode } from '@/context/DarkModeContext';
 import MoonIcon from '../icons/MoonIcon';
-import styles from './css/ThemeButton.module.css';
+import styles from './css/index.module.css';
 
 export default function ThemeButton() {
   const { darkMode, toggleDarkMode } = useDarkMode();
